Databasefinal

+++ app/src/main/java/com/example/suggestion/data/ExpenseDao.kt
@@ -13,7 +13,7 @@ interface ExpenseDao {
     suspend fun insertExpense(expense: Expense): Long
 
     @Query("SELECT * FROM expenses WHERE userId = :userId")
-    suspend fun getExpensesForUser(userId: Int): List<Expense>
+    suspend fun getExpensesForUser(userId: Long): List<Expense>
 
     @Query("SELECT * FROM expenses")
     suspend fun getAllExpenses(): List<Expense>
